+++ eResourceWeb/Models/ResourceMaster.cs
@@ -9,22 +9,22 @@ namespace eResourceWeb.Models
     public class ResourceMaster : BaseModel
     {
         [Key]
-        public float ResourceId { get; set; }
+        public Double ResourceId { get; set; }
         public string ResourceName { get; set; }
-        public int TypeId { get; set; }
-        public int ManagerId { get; set; }
-        public int Status { get; set; }
-        public int DesignationId { get; set; }
-        public int RegionId { get; set; }
-        public int OfficeId { get; set; }
-        public int DepartmentId { get; set; }
-        public int TeamId { get; set; }
+        public Double TypeId { get; set; }
+        public Double ManagerId { get; set; }
+        public Double Status { get; set; }
+        public Double DesignationId { get; set; }
+        public Double RegionId { get; set; }
+        public Double OfficeId { get; set; }
+        public Double DepartmentId { get; set; }
+        public Double TeamId { get; set; }
         public string Email { get; set; }
         public string Notes { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime TerminationDate { get; set; }
         public string Phone { get; set; }
-        public int TimeSheetApproval { get; set; }
+        public Double TimeSheetApproval { get; set; }
         public string UpdatedBy { get; set; }
         public DateTime UpdateDate { get; set; }
